arc: addons/extensions: fix code from codefactor

+++ files/board/arpl/overlayfs/opt/arpl/include/addons.sh
@@ -64,7 +64,7 @@ function installAddon() {
   cp "${TMP_PATH}/${ADDON}/install.sh" "${RAMDISK_PATH}/addons/${ADDON}.sh" 2>"${LOG_FILE}" || dieLog
   chmod +x "${RAMDISK_PATH}/addons/${ADDON}.sh"
   [ -d ${TMP_PATH}/${ADDON}/root ] && (cp -R "${TMP_PATH}/${ADDON}/root/"* "${RAMDISK_PATH}/" 2>"${LOG_FILE}" || dieLog)
-  rm -rf "${TMP_PATH}/${ADDON}"
+  rm -rf "${TMP_PATH}/${ADDON:?}"
   return 0
 }
 
@@ -73,12 +73,12 @@ function installAddon() {
 # 1 - Addon file path
 # Return name of addon on sucess or empty on error
 function untarAddon() {
-  rm -rf "${TMP_PATH}/addon"
-  mkdir -p "${TMP_PATH}/addon"
-  tar -xaf "${1}" -C "${TMP_PATH}/addon" || return
-  ADDON=$(readConfigKey "name" "${TMP_PATH}/addon/manifest.yml")
+  rm -rf "${TMP_PATH}/${ADDON:?}"
+  mkdir -p "${TMP_PATH}/${ADDON}"
+  tar -xaf "${1}" -C "${TMP_PATH}/${ADDON}" || return
+  ADDON=$(readConfigKey "name" "${TMP_PATH}/${ADDON}/manifest.yml")
   [ -z "${ADDON}" ] && return
-  rm -rf "${ADDONS_PATH}/${ADDON}"
-  mv "${TMP_PATH}/addon" "${ADDONS_PATH}/${ADDON}"
+  rm -rf "${ADDONS_PATH}/${ADDON:?}"
+  mv "${TMP_PATH}/${ADDON}" "${ADDONS_PATH}/${ADDON}"
   echo "${ADDON}"
 }
